PerformOCR/requirements.txt
pytesseract
Pillow
pika
python-multipart